Somewhat fixed bug
So the problem is that on teleport, the player object is disable, moved, then re-enabled. Disabling an object doesn't trigger OnTriggerExit, because unity. This will come down to Bill to figure out how to handle, as it could cause problems for other people's OnTriggerExits as well.
At the very least, I did fix a problem with the pressure plates. Now, if two things are on the same pressure plate, it won't activate it multiple times